# spark-ai-server/src/main/resources/openapi/dynamic-data-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SPARK Server API",
    "version": "1.0.0",
    "description": "Complete SPARK server API documentation, including authentication, app config, page config, AI, projects, navigation, cache, diagnostics, metadata-driven dynamic tables, unified CRUD, DataView query, tree operations, batch jobs, and SSE notifications."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/tenants/{tenantId}/projects/{projectId}/data/{tableName}/query": {
      "post": {
        "tags": [
          "Dynamic Data"
        ],
        "summary": "Query a metadata table using DataView config",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/TableName"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataViewQueryRequest"
              },
              "examples": {
                "ordersDataView": {
                  "$ref": "#/components/examples/OrdersDataViewQuery"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rows, paging, and optional aggregate result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataViewQueryResult"
                }
              }
            }
          }
        }
      }
    },
    "/tenants/{tenantId}/projects/{projectId}/data/{tableName}/records": {
      "post": {
        "tags": [
          "Dynamic Data"
        ],
        "summary": "Create one record",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/TableName"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataRecord"
              },
              "examples": {
                "orderRecord": {
                  "$ref": "#/components/examples/CreateOrderRecord"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataRecord"
                }
              }
            }
          }
        }
      }
    },
    "/tenants/{tenantId}/projects/{projectId}/data/{tableName}/records/get": {
      "post": {
        "tags": [
          "Dynamic Data"
        ],
        "summary": "Get one record by primary key",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/TableName"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrimaryKeyPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataRecord"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/tenants/{tenantId}/projects/{projectId}/data/{tableName}/records/update": {
      "post": {
        "tags": [
          "Dynamic Data"
        ],
        "summary": "Update one record. Body includes primary key fields plus patch fields.",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/TableName"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataRecord"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataRecord"
                }
              }
            }
          }
        }
      }
    },
    "/tenants/{tenantId}/projects/{projectId}/data/{tableName}/records/delete": {
      "post": {
        "tags": [
          "Dynamic Data"
        ],
        "summary": "Delete one record by primary key",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/TableName"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrimaryKeyPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          }
        }
      }
    },
    "/tenants/{tenantId}/projects/{projectId}/data/{tableName}/records/batch-create": {
      "post": {
        "tags": [
          "Dynamic Data"
        ],
        "summary": "Batch create records",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/TableName"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/RecordArray"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/BatchResult"
          }
        }
      }
    },
    "/tenants/{tenantId}/projects/{projectId}/data/{tableName}/records/batch-update": {
      "post": {
        "tags": [
          "Dynamic Data"
        ],
        "summary": "Batch update records",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/TableName"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/RecordArray"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/BatchResult"
          }
        }
      }
    },
    "/tenants/{tenantId}/projects/{projectId}/data/{tableName}/records/batch-delete": {
      "post": {
        "tags": [
          "Dynamic Data"
        ],
        "summary": "Batch delete records by primary keys",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/TableName"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/RecordArray"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/BatchResult"
          }
        }
      }
    },
    "/tenants/{tenantId}/projects/{projectId}/data/{tableName}/tree/children": {
      "post": {
        "tags": [
          "Dynamic Tree"
        ],
        "summary": "Load direct children",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/TableName"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TreeChildrenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Flat tree nodes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlatTreeNode"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tenants/{tenantId}/projects/{projectId}/data/{tableName}/tree/path": {
      "post": {
        "tags": [
          "Dynamic Tree"
        ],
        "summary": "Load ancestor path IDs",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/TableName"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TreePathRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Path IDs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreePathResult"
                }
              }
            }
          }
        }
      }
    },
    "/tenants/{tenantId}/projects/{projectId}/data/{tableName}/tree/subtree": {
      "post": {
        "tags": [
          "Dynamic Tree"
        ],
        "summary": "Load path/subtree nodes for expand-to-node",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/TableName"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TreeSubtreeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Node map keyed by node id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/FlatTreeNode"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tenants/{tenantId}/projects/{projectId}/data/{tableName}/tree/move": {
      "post": {
        "tags": [
          "Dynamic Tree"
        ],
        "summary": "Move a node to a new parent",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/TableName"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TreeMoveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Moved node",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "node": {
                      "$ref": "#/components/schemas/FlatTreeNode"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tenants/{tenantId}/projects/{projectId}/data/{tableName}/tree/search": {
      "post": {
        "tags": [
          "Dynamic Tree"
        ],
        "summary": "Search flat tree nodes with ancestor path",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/TableName"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/TreeSearch"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/NestedTreeSearchResult"
          }
        }
      }
    },
    "/tenants/{tenantId}/projects/{projectId}/data/{tableName}/tree/nested": {
      "post": {
        "tags": [
          "Dynamic Tree"
        ],
        "summary": "Load nested tree",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/TableName"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NestedTreeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Nested tree nodes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NestedTreeNode"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tenants/{tenantId}/projects/{projectId}/data/{tableName}/tree/nested/search": {
      "post": {
        "tags": [
          "Dynamic Tree"
        ],
        "summary": "Search nested tree nodes with ancestor path",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/TableName"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/TreeSearch"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/NestedTreeSearchResult"
          }
        }
      }
    },
    "/tenants/{tenantId}/projects/{projectId}/data/transactions": {
      "post": {
        "tags": [
          "Dynamic Data"
        ],
        "summary": "Commit multi-table operations in one transaction",
        "description": "Executes all operations synchronously in one backing database transaction. Any failed operation rolls back the whole transaction. requestId is an idempotency key shared through the database, so retries routed to another backend instance replay the committed result instead of executing the operations again. data-change SSE events are emitted only after commit by the current in-process SSE service; multi-instance deployments need an outbox or message broker for cross-instance event fan-out.",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionRequest"
              },
              "examples": {
                "multiTable": {
                  "$ref": "#/components/examples/MultiTableTransaction"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Committed transaction result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/tenants/{tenantId}/projects/{projectId}/data/batch-jobs": {
      "post": {
        "tags": [
          "Batch Jobs"
        ],
        "summary": "Submit an asynchronous multi-table update job",
        "description": "Progress and completion are emitted through SSE event data-batch-job. Data changes are emitted through data-change.",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchJobRequest"
              },
              "examples": {
                "multiTable": {
                  "$ref": "#/components/examples/MultiTableBatchJob"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Queued job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchJobQueued"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "TenantId": {
        "name": "tenantId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "example": "lmspark"
      },
      "ProjectId": {
        "name": "projectId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "example": "default"
      },
      "TableName": {
        "name": "tableName",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[A-Za-z_][A-Za-z0-9_]*$"
        },
        "example": "Orders"
      },
...
    },
    "requestBodies": {
...
      "RecordArray": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/DataRecord"
              }
            }
          }
        }
      },
      "TreeSearch": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TreeSearchRequest"
            }
          }
        }
      }
    },
    "responses": {
...
      "Ok": {
        "description": "OK",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "ok": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Not found"
      },
      "BatchResult": {
        "description": "Batch operation result",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BatchResult"
            }
          }
        }
      },
      "NestedTreeSearchResult": {
        "description": "Search results with path",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/NestedTreeSearchResult"
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "DataViewQueryRequest": {
        "type": "object",
        "properties": {
          "viewId": {
            "type": "string"
          },
          "viewConfig": {
            "$ref": "#/components/schemas/DataViewConfig"
          },
          "query": {
            "$ref": "#/components/schemas/QueryParams"
          }
        }
      },
...
      "DataViewQueryResult": {
        "type": "object",
        "properties": {
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataRecord"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "aggregateResult": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "DataRecord": {
        "type": "object",
        "additionalProperties": true,
        "example": {
          "id": 1,
          "name": "Order A",
          "status": "open",
          "amount": 128.5
        }
      },
      "PrimaryKeyPayload": {
        "type": "object",
        "additionalProperties": true,
        "example": {
          "id": 1
        }
      },
...
      "BatchJobQueued": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "example": "queued"
          }
        }
      },
...
      "FlatTreeNode": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "id": {},
          "parentId": {},
          "name": {
            "type": "string"
          },
          "hasChildren": {
            "type": "boolean"
          }
        }
      },
...
      "TreeChildrenRequest": {
        "type": "object",
        "properties": {
          "parentId": {},
          "limit": {
            "type": "integer"
          },
          "treeMode": {
            "type": "string",
            "enum": [
              "flat",
              "nested"
            ]
          },
          "viewConfig": {
            "$ref": "#/components/schemas/DataViewConfig"
          }
        }
      },
      "TreePathRequest": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {}
        }
      },
      "TreeSubtreeRequest": {
        "type": "object",
        "required": [
          "toId"
        ],
        "properties": {
          "fromId": {},
          "toId": {},
          "includeTargetChildren": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "TreeMoveRequest": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {},
          "newParentId": {},
          "index": {
            "type": "integer"
          }
        }
      },
...
      "TreePathResult": {
        "type": "object",
        "properties": {
          "pathIds": {
            "type": "array",
            "items": {}
          }
        }
      },
...
    },
    "examples": {
...
      "CreateOrderRecord": {
        "summary": "Create one order",
        "value": {
          "name": "Order A",
          "status": "open",
          "amount": 128.5,
          "parentId": null
        }
      },
      "OrdersDataViewQuery": {
        "summary": "DataView filter, sort, pagination, tree config, aggregate",
        "value": {
          "viewId": "default",
          "query": {
            "page": 1,
            "pageSize": 10,
            "sort": "amount:desc",
            "filter": {
              "type": "and",
              "children": [
                {
                  "field": "status",
                  "op": "==",
                  "value": "open"
                },
                {
                  "field": "amount",
                  "op": ">=",
                  "value": 10
                }
              ]
            },
            "search": "Order",
            "treeMode": "flat"
          },
          "viewConfig": {
            "treeConfig": {
              "idField": "id",
              "parentIdField": "parentId",
              "textField": "name",
              "treeMode": "flat"
            },
            "aggregates": {
              "totalAmount": {
                "type": "sum",
                "field": "amount"
              }
            }
          }
        }
      },
...
    }
  }
}